Move unmodifiable code to module.
